Separate item search selectors from common helpers
