+++ src/app/dim-api/reducer.ts
@@ -16,7 +16,7 @@ import { convertDimLoadoutToApiLoadout } from 'app/loadout-drawer/loadout-type-c
 import { recentSearchComparator } from 'app/search/autocomplete';
 import { CUSTOM_TOTAL_STAT_HASH } from 'app/search/d2-known-values';
 import { FilterContext } from 'app/search/filter-types';
-import { buildItemFiltersMap } from 'app/search/search-config';
+import { buildItemFiltersMap } from 'app/search/item-search-config';
 import { parseAndValidateQuery } from 'app/search/search-utils';
 import { count, uniqBy } from 'app/utils/collections';
 import { emptyArray } from 'app/utils/empty';
