+++ src/app/inventory/ConnectedInventoryItem.tsx
@@ -1,8 +1,8 @@
 import { settingSelector } from 'app/dim-api/selectors';
+import { queryValidSelector, searchFilterSelector } from 'app/search/item-search-filter';
 import _ from 'lodash';
 import React, { useMemo } from 'react';
 import { useSelector } from 'react-redux';
-import { queryValidSelector, searchFilterSelector } from '../search/search-filter';
 import { wishListSelector } from '../wishlists/selectors';
 import InventoryItem from './InventoryItem';
 import { DimItem } from './item-types';
