+++ src/app/item-actions/ItemActionsDropdown.tsx
@@ -10,8 +10,8 @@ import { DimStore } from 'app/inventory/store-types';
 import { itemMoveLoadout } from 'app/loadout-drawer/auto-loadouts';
 import { applyLoadout } from 'app/loadout-drawer/loadout-apply';
 import { TagCommandInfo } from 'app/organizer/ItemActions';
+import { validateQuerySelector } from 'app/search/item-search-filter';
 import { canonicalizeQuery, parseQuery } from 'app/search/query-parser';
-import { validateQuerySelector } from 'app/search/search-filter';
 import { toggleSearchResults } from 'app/shell/actions';
 import { useIsPhonePortrait } from 'app/shell/selectors';
 import { useThunkDispatch } from 'app/store/thunk-dispatch';
