+++ src/app/progress/Pursuit.tsx
@@ -5,7 +5,7 @@ import { isNewSelector } from 'app/inventory/selectors';
 import { isBooleanObjective } from 'app/inventory/store/objectives';
 import ItemExpiration from 'app/item-popup/ItemExpiration';
 import { useD2Definitions } from 'app/manifest/selectors';
-import { searchFilterSelector } from 'app/search/search-filter';
+import { searchFilterSelector } from 'app/search/item-search-filter';
 import { percent } from 'app/shell/formatters';
 import { RootState } from 'app/store/types';
 import clsx from 'clsx';
